@@ -1,16 +1,17 @@
 {
 	"name": "@szmarczak/http-timer",
-	"version": "4.0.6",
+	"version": "5.0.0",
 	"description": "Timings for HTTP requests",
-	"main": "dist/source",
+	"type": "module",
+	"exports": "./dist/source/index.js",
 	"engines": {
-		"node": ">=10"
+		"node": ">=14.16"
 	},
 	"scripts": {
-		"test": "xo && tsc --noEmit && nyc ava",
+		"test": "xo && ava",
 		"build": "del-cli dist && tsc",
 		"prepare": "npm run build",
-		"coveralls": "nyc report --reporter=text-lcov | coveralls"
+		"coveralls": "exit 0 && nyc report --reporter=text-lcov | coveralls"
 	},
 	"files": [
 		"dist/source"
@@ -18,8 +19,11 @@
 	"keywords": [
 		"http",
 		"https",
+		"http2",
 		"timer",
-		"timings"
+		"timings",
+		"performance",
+		"measure"
 	],
 	"repository": {
 		"type": "git",
@@ -32,20 +36,21 @@
 	},
 	"homepage": "https://github.com/szmarczak/http-timer#readme",
 	"dependencies": {
-		"defer-to-connect": "^2.0.0"
+		"defer-to-connect": "^2.0.1"
 	},
 	"devDependencies": {
 		"@ava/typescript": "^2.0.0",
 		"@sindresorhus/tsconfig": "^1.0.2",
-		"@types/node": "^16.3.1",
+		"@types/node": "^16.7.0",
 		"ava": "^3.15.0",
 		"coveralls": "^3.1.1",
-		"del-cli": "^3.0.1",
-		"http2-wrapper": "^2.0.7",
+		"del-cli": "^4.0.1",
+		"http2-wrapper": "^2.1.4",
 		"nyc": "^15.1.0",
 		"p-event": "^4.2.0",
+		"ts-node": "^10.2.1",
 		"typescript": "^4.3.5",
-		"xo": "^0.39.1"
+		"xo": "^0.44.0"
 	},
 	"types": "dist/source",
 	"nyc": {
@@ -58,15 +63,26 @@
 	},
 	"xo": {
 		"rules": {
-			"@typescript-eslint/no-non-null-assertion": "off"
+			"@typescript-eslint/no-non-null-assertion": "off",
+			"@typescript-eslint/no-unsafe-assignment": "off",
+			"@typescript-eslint/no-unsafe-member-access": "off",
+			"@typescript-eslint/no-unsafe-call": "off"
 		}
 	},
 	"ava": {
-		"typescript": {
-			"compile": false,
-			"rewritePaths": {
-				"tests/": "dist/tests/"
-			}
-		}
+		"files": [
+			"tests/*"
+		],
+		"timeout": "1m",
+		"nonSemVerExperiments": {
+			"nextGenConfig": true,
+			"configurableModuleFormat": true
+		},
+		"extensions": {
+			"ts": "module"
+		},
+		"nodeArguments": [
+			"--loader=ts-node/esm"
+		]
 	}
 }
